spd-diplom/social_network/requirements.txt
django==5.0.2
djangorestframework==3.14.0
pillow==10.2.0
psycopg2-binary==2.9.9
geopy==2.4.1
python-dotenv==1.0.1
flake8==7.1.0
black==24.8.0
